+++ stdlib/LinearAlgebra/src/LinearAlgebra.jl
@@ -126,7 +126,6 @@ export
     qr,
     qrfact!,
     qrfact,
-    lq,
     lqfact!,
     lqfact,
     rank,